requirements.txt
Django==3.2.6
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
python-environ==0.4.54
psycopg2-binary==2.9.1
